// src/routes/api/youtube/types.ts
export type ITEM = {
	id?: s
	type?: s
	thumbnail?: s
	title?: s
	length?: s
	videos?: any
	videoCount?: s
	isLive?: b
	channelTitle?: s
	shortBylineText?: s
}